% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-route.R
\name{correct_distances}
\alias{correct_distances}
\title{Correct the elevation distances}
\usage{
correct_distances(dists)
}
\arguments{
\item{dists}{numeric from the elevation first column}
}
\description{
OTP returns elevation as a distance along the leg, resetting to 0 at each leg
but we need the distance along the total route. so calculate this
}
