% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json2sf.R
\name{json2sf_otp}
\alias{json2sf_otp}
\title{Convert output from Open Trip Planner into sf object}
\usage{
json2sf_otp(obj)
}
\arguments{
\item{obj}{Object from transportapi.com read-in with}
}
\description{
Convert output from Open Trip Planner into sf object
}
\examples{
None
}
