% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{json_example_long_drive}
\alias{json_example_long_drive}
\title{Example JSON for driving long distance}
\format{
json
}
\usage{
json_example_long_drive
}
\description{
Example JSON response from OTP
This is used for internal testing and has no use
}
\keyword{datasets}
