% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{json_example_transit}
\alias{json_example_transit}
\title{Example JSON for transit}
\format{
json
}
\usage{
json_example_transit
}
\description{
Example JSON response from OTP
This is used for internal testing and has no use
}
\keyword{datasets}
