% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentripplanner-package.R
\docType{package}
\name{opentripplanner-package}
\alias{opentripplanner}
\alias{opentripplanner-package}
\title{Open Trip Planner of R}
\description{
The goal of Open Trip Planner for R is to provide a simple R interface to Open Trip Planner (OTP).
The OTP is a multimodal trip planning service. OTP is written in Java, and you will need both Java
and R to use this package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ITSLeeds/opentripplanner}
  \item \url{https://itsleeds.github.io/opentripplanner/}
  \item Report bugs at \url{https://github.com/ITSLeeds/opentripplanner/issues}
}

}
\author{
\strong{Maintainer}: Marcus Young \email{M.A.Young@soton.ac.uk} (0000-0003-4627-1116)

Authors:
\itemize{
  \item Malcolm Morgan \email{m.morgan1@leeds.ac.uk} (0000-0002-9488-9183)
}

Other contributors:
\itemize{
  \item Robin Lovelace \email{rob00x@gmail.com} (0000-0001-5679-6536) [contributor]
  \item Layik Hama \email{layik.hama@gmail.com} (0000-0003-1912-4890) [contributor]
}

}
\keyword{mulitmodal}
\keyword{opentripplanner}
\keyword{routing}
\keyword{transport}
