% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-plan.R
\name{otp_async}
\alias{otp_async}
\title{Async Send  Requests}
\usage{
otp_async(urls, ncores)
}
\arguments{
\item{urls}{vector of URLs for OTP}

\item{ncores}{Number of requests to send at once}
}
\description{
Async Send  Requests
}
\concept{internal}
