% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-setup.R
\name{otp_check_java}
\alias{otp_check_java}
\title{Check Java version}
\usage{
otp_check_java(otp_version = 1.5)
}
\arguments{
\item{otp_version}{numeric, OTP version number default 1.5}
}
\description{
Check if you have the correct version of Java for running OTP locally
}
\seealso{
Other setup: 
\code{\link{otp_build_graph}()},
\code{\link{otp_check_version}()},
\code{\link{otp_dl_demo}()},
\code{\link{otp_dl_jar}()},
\code{\link{otp_make_config}()},
\code{\link{otp_setup}()},
\code{\link{otp_stop}()},
\code{\link{otp_validate_config}()},
\code{\link{otp_write_config}()}
}
\concept{setup}
