% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-connect.R
\name{otp_check_version}
\alias{otp_check_version}
\title{Check the what version of OTP the server is running}
\usage{
otp_check_version(otpcon, warn = TRUE)
}
\arguments{
\item{otpcon}{otpcon object from otp_connect()}

\item{warn}{logical, if TRUE will check that OTP version matches contents of otpcon}
}
\description{
Check the what version of OTP the server is running
}
\seealso{
Other setup: 
\code{\link{otp_build_graph}()},
\code{\link{otp_check_java}()},
\code{\link{otp_dl_demo}()},
\code{\link{otp_dl_jar}()},
\code{\link{otp_make_config}()},
\code{\link{otp_setup}()},
\code{\link{otp_stop}()},
\code{\link{otp_validate_config}()},
\code{\link{otp_write_config}()}
}
\concept{setup}
