% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-setup.R
\name{otp_checks}
\alias{otp_checks}
\title{Basic OTP Setup Checks}
\usage{
otp_checks(otp = NULL, dir = NULL, router = NULL, graph = FALSE)
}
\arguments{
\item{dir}{A character string path to a folder containing the necessary files, see details}

\item{router}{A character string for the name of the router, must match with contents of dir, default "current"}

\item{graph}{Logical, check for graph, default = FALSE}
}
\description{
Checks to run before setting up the OTP
}
