% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-plan.R
\name{otp_clean_input}
\alias{otp_clean_input}
\title{Clean Batch Inputs}
\usage{
otp_clean_input(imp, imp_name)
}
\arguments{
\item{imp}{fromPlace or toPlace input}

\item{imp_name}{name of input}
}
\description{
Clean numeric, SF, or matrix prior to routing
}
