% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-connect.R
\name{otp_connect}
\alias{otp_connect}
\title{Set up and confirm a connection to an OTP instance.}
\usage{
otp_connect(hostname = "localhost", router = "default", port = 8080,
  ssl = FALSE)
}
\arguments{
\item{hostname}{A string, e.g. "ec2-34-217-73-26.us-west-2.compute.amazonaws.com".
Default is "localhost".}

\item{router}{A string, e.g. "UK2018". Default is "default".}

\item{port}{A positive integer. Default is 8080.}

\item{ssl}{A logical value, indicates whether to use https. Default is FALSE.}
}
\value{
Returns an S3 object of class otpconnect. This object is entered as
an argument in the functions that are used to query an OTP instance.
}
\description{
Defines the parameters required to connect to a router on an OTP instance and
confirms that the instance is queryable.
}
\examples{
\dontrun{
otpcon <- otp_connect()
otpcon <- otp_connect(router = "UK-2018",
                      ssl = TRUE)
otpcon <- otp_connect(hostname = "ec2.us-west-2.compute.amazonaws.com",
                      router = "UK2018",
                      port = 8888,
                      ssl = TRUE)
}
}
