% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-download.R
\name{otp_dl_demo}
\alias{otp_dl_demo}
\title{Download Demo Data}
\usage{
otp_dl_demo(
  path_data = NULL,
  url = paste0("https://github.com/ropensci/opentripplanner/",
    "releases/download/0.1/isle-of-wight-demo.zip")
)
}
\arguments{
\item{path_data}{path to folder where data for OTP is to be stored}

\item{url}{URL to data}
}
\description{
Download the demonstration data for the Isle of Wight
}
\examples{
\dontrun{
otp_dl_demo(tempdir())
}
}
\seealso{
Other setup: 
\code{\link{otp_build_graph}()},
\code{\link{otp_dl_jar}()},
\code{\link{otp_make_config}()},
\code{\link{otp_setup}()},
\code{\link{otp_stop}()},
\code{\link{otp_validate_config}()},
\code{\link{otp_write_config}()}
}
\concept{setup}
