% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-download.R
\name{otp_dl_jar}
\alias{otp_dl_jar}
\title{Download OTP Jar File}
\usage{
otp_dl_jar(
  path = NULL,
  version = "1.4.0",
  file_name = "otp.jar",
  url = "https://repo1.maven.org/maven2/org/opentripplanner/otp"
)
}
\arguments{
\item{path}{path to folder where OTP is to be stored}

\item{version}{a character string of the version number default is "1.4.0"}

\item{file_name}{file name to give the otp default "otp.jar"}

\item{url}{URL to the dowload server}
}
\value{
The path to the OTP file
}
\description{
Download the OTP jar file from maven.org
}
\examples{
\dontrun{
otp_dl_jar(tempdir())
}
}
\seealso{
Other setup: 
\code{\link{otp_build_graph}()},
\code{\link{otp_dl_demo}()},
\code{\link{otp_make_config}()},
\code{\link{otp_setup}()},
\code{\link{otp_stop}()},
\code{\link{otp_validate_config}()},
\code{\link{otp_write_config}()}
}
\concept{setup}
