% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-plan.R
\name{otp_get_results}
\alias{otp_get_results}
\title{Get OTP results}
\usage{
otp_get_results(x, otpcon, fromPlace, toPlace, ... = ...)
}
\arguments{
\item{x}{numeric}

\item{otpcon}{otpcon}

\item{fromPlace}{fromplace}

\item{toPlace}{toPlace}

\item{...}{all other variaibles}
}
\description{
helper function for otp_plan
}
