% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-route.R
\name{otp_json2sf}
\alias{otp_json2sf}
\title{Convert output from Open Trip Planner into sf object}
\usage{
otp_json2sf(obj, full_elevation = FALSE)
}
\arguments{
\item{obj}{Object from the OTP API to process}

\item{full_elevation}{logical should the full elevation profile be returned (if available)}
}
\description{
Convert output from Open Trip Planner into sf object
}
