% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-config.R
\name{otp_make_config}
\alias{otp_make_config}
\title{Make Config Object}
\usage{
otp_make_config(type)
}
\arguments{
\item{type}{Which type of config file to create, "otp", "build", "router"}
}
\description{
OTP can be configured using three json files `otp-config.json`, `build-config.json`,
and `router-config.json`. This function creates a named list for each config file and
populates the defaults values.
}
\details{
For more details see:
http://docs.opentripplanner.org/en/latest/Configuration
}
