% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-batch.R
\name{otp_plan_batch}
\alias{otp_plan_batch}
\title{Get multiple routes from the OTP}
\usage{
otp_plan_batch(otpcon = NA, fromPlace = NA, toPlace = NA,
  ncores = 1, ...)
}
\arguments{
\item{otpcon}{OTP connection object produced by otp_connect()}

\item{fromPlace}{Numeric matrix of two columns, Latitude/Longitude pairs}

\item{toPlace}{Numeric matrix of two columns, Latitude/Longitude pairs}

\item{ncores}{numeric, number of cores to use in parallel processing, default is 1}

\item{...}{Other variables passed to otp_plan}
}
\value{
Returns a data.frame of SF POLYLINES
}
\description{
Get multiple routes from the OTP
}
\details{
This function is a batch version of otp_plan() and is useful if you want to produce many routes at once.
}
