% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-plan.R
\name{otp_plan_internal}
\alias{otp_plan_internal}
\title{Get the geometry of a route from the OTP}
\usage{
otp_plan_internal(otpcon = NA, fromPlace = NA, toPlace = NA,
  mode = "CAR", date = date, time = time, arriveBy = FALSE,
  maxWalkDistance = 1000, walkReluctance = 2, transferPenalty = 0,
  minTransferTime = 0, numItineraries = 3, full_elevation = FALSE,
  get_geometry = TRUE)
}
\arguments{
\item{otpcon}{OTP connection object produced by otp_connect()}

\item{fromPlace}{Numeric vector, Latitude/Longitude pair, e.g. `c(51.529258,-0.134649)`}

\item{toPlace}{Numeric vector, Latitude/Longitude pair, e.g. `c(51.506383,-0.088780,)`}

\item{mode}{Character vector of modes of travel valid values TRANSIT, WALK, BICYCLE, CAR, BUS, RAIL, default CAR}

\item{date}{date}

\item{time}{time}

\item{arriveBy}{Logical, Whether the trip should depart or arrive at the specified date and time, default FALSE}

\item{maxWalkDistance}{Numeric passed to OTP}

\item{walkReluctance}{Numeric passed to OTP}

\item{transferPenalty}{Numeric passed to OTP}

\item{minTransferTime}{Numeric passed to OTP}

\item{numItineraries}{The maximum number of possible itineraries to return}

\item{full_elevation}{Logical, should the full elevation profile be returned, default FALSE}
}
\description{
Get the geometry of a route from the OTP
}
\details{
This function returns a SF data.frame with one row for each leg of the journey
(a leg is defined by a change in mode). For transit more than one route option may be returned
and is indicated by the route_option column.
}
