% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-surface.R
\name{otp_pointset}
\alias{otp_pointset}
\title{Create a pointset}
\usage{
otp_pointset(points = NULL, name = NULL, dir = NULL)
}
\arguments{
\item{points}{sf data frame of POINTS with CRS 4326}

\item{name}{Character, name for pointset}

\item{dir}{A character string, path to a directory containing the necessary
files, see details}
}
\value{
Returns a data.frame of SF POINTS or Coordinates of all
    the locations that match `query`
}
\description{
Pointsets are text files tha can be used by the Analyist feature in OTP 1.5
}
\details{
OTP will return a maximum of 10 results
}
\examples{
\dontrun{
locations <- otp_geocode(otpcon, "High Street")
}
}
\seealso{
Other routing: 
\code{\link{otp_geocode}()},
\code{\link{otp_isochrone}()},
\code{\link{otp_plan}()},
\code{\link{otp_routing_options}()},
\code{\link{otp_validate_routing_options}()}
}
\concept{routing}
