% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-surface.R
\name{otp_surface}
\alias{otp_surface}
\title{Use OTP Geo-coder to find a location}
\usage{
otp_surface(
  otpcon = NULL,
  fromPlace = NULL,
  pointsset = NULL,
  mode = "CAR",
  date_time = Sys.time(),
  maxWalkDistance = 800,
  arriveBy = FALSE,
  timezone = otpcon$timezone
)
}
\arguments{
\item{otpcon}{OTP connection object produced by otp_connect()}

\item{fromPlace}{Numeric vector, Longitude/Latitude pair, e.g.
`c(-0.134649,51.529258)`, or 2 column matrix of Longitude/Latitude pairs,
or sf data frame of POINTS with CRS 4326}

\item{pointsset}{character, name of pointset}

\item{date_time}{POSIXct, a date and time, defaults to current date and time}

\item{maxWalkDistance}{Numeric passed to OTP in metres}

\item{arriveBy}{Logical, Whether the trip should depart or arrive at the
specified date and time, default FALSE}

\item{timezone}{Character, what timezone to use, see as.POSIXct, default is
local timezone}

\item{dir}{A character string, path to a directory containing the necessary
files, see details}
}
\value{
Returns a data.frame of travel times
}
\description{
Geo-coding converts a named place, such as a street name into a lng/lat pair.
}
\details{
THis function requires the analysis and pointset features to be
enabled during `otp_setup()`. Thus it will only work with OTP 1.x. For more
detail see the analyst vignette.
}
\examples{
\dontrun{
times <- otp_surface(otpcon, c(-1.17502, 50.64590), "lsoa", path_data)
}
}
\concept{analyst}
