% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-config.R
\name{otp_validate_config}
\alias{otp_validate_config}
\title{Validate Config Object}
\usage{
otp_validate_config(config, type = attributes(config)$config_type)
}
\arguments{
\item{config}{A named list made/modified from `otp_make_config()`}
}
\description{
Checks if the list of OTP configiration options is valid
}
\details{
Performs basic validity checks on class, max/min values etc as appropiate, some of
more compex parameters are not checked. For more details see:

http://docs.opentripplanner.org/en/latest/Configuration
http://dev.opentripplanner.org/javadoc/1.3.0/org/opentripplanner/routing/core/RoutingRequest.html
}
