% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isochrone.R
\name{otpIsochrone}
\alias{otpIsochrone}
\title{Plan a journey with Open Trip Planner}
\usage{
otpIsochrone(otpcon, from, modes, cutoff, batch = TRUE, date = "2017/06/12",
  time = "09:00:00")
}
\arguments{
\item{from}{Longitude/Latitude pair, e.g. \code{c(-0.134649,51.529258)} or SF points of class "sfc_POINT" "sfc" and length one}

\item{to}{Longitude/Latitude pair, e.g. \code{c(-0.088780,51.506383)} or SF points of class "sfc_POINT" "sfc" and length one}

\item{save_raw}{Boolean value which returns raw list from the geojson if TRUE (FALSE by default).}
}
\description{
R interface to the Open Trip Planner journey planning API,
See \href{http://docs.opentripplanner.org/en/latest/}{docs.opentripplanner.org} for details.
}
\details{
Requires a local instance of Open Trip Planner.
See the OTP \href{http://docs.opentripplanner.org/en/latest/Basic-Tutorial/}{tutorials} for set up details.
}
\examples{
\dontrun{
from = c(-0.134649,51.529258) # Euston Station
to = c(-0.088780,51.506383) # Bridge House
r1 = journey(from, to)
}
}
\seealso{
json2sf_tapi
}
