% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-route.R
\name{polyline2linestring}
\alias{polyline2linestring}
\title{Convert Google Encoded Polyline and elevation data into sf object}
\usage{
polyline2linestring(line, elevation = NULL)
}
\arguments{
\item{line}{character - polyline}

\item{elevation}{numeric - vector of elevations
None}
}
\description{
OTP returns the 2d route as a polylinebean and the elevation profile as vector of numbers
But the number of points for each is not the same, as 2D line only has a point at change of driections
While elevation is regually spaced. If elevation is supplied the correct heights are matched
}
