% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opentripplanner-package.R
\docType{package}
\name{opentripplanner-package}
\alias{opentripplanner}
\alias{opentripplanner-package}
\title{OpenTripPlanner of R}
\description{
The goal of OpenTripPlanner for R is to provide a simple R interface
to OpenTripPlanner (OTP).
The OTP is a multimodal trip planning service.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/ropensci/opentripplanner}
  \item \url{https://docs.ropensci.org/opentripplanner/}
  \item Report bugs at \url{https://github.com/ropensci/opentripplanner/issues}
}

}
\author{
\strong{Maintainer}: Malcolm Morgan \email{m.morgan1@leeds.ac.uk} (\href{https://orcid.org/0000-0002-9488-9183}{ORCID})

Authors:
\itemize{
  \item Marcus Young \email{M.A.Young@soton.ac.uk} (\href{https://orcid.org/0000-0003-4627-1116}{ORCID})
  \item Robin Lovelace \email{rob00x@gmail.com} (\href{https://orcid.org/0000-0001-5679-6536}{ORCID})
}

Other contributors:
\itemize{
  \item Layik Hama \email{layik.hama@gmail.com} (\href{https://orcid.org/0000-0003-1912-4890}{ORCID}) [contributor]
}

}
\keyword{mulitmodal}
\keyword{opentripplanner}
\keyword{routing}
\keyword{transport}
