% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-route.R
\name{otp_plan}
\alias{otp_plan}
\title{Get the geometry of a route from the OTP}
\usage{
otp_plan(otpcon = NA, fromPlace = NA, toPlace = NA, mode = "CAR",
  date_time = Sys.time(), arriveBy = FALSE, maxWalkDistance = 800,
  walkReluctance = 2, transferPenalty = 0, minTransferTime = 0,
  full_elevation = FALSE)
}
\arguments{
\item{otpcon}{OTP connection object produced by otp_connect()}

\item{fromPlace}{Numeric vector, Latitude/Longitude pair, e.g. `c(51.529258,-0.134649)`}

\item{toPlace}{Numeric vector, Latitude/Longitude pair, e.g. `c(51.506383,-0.088780,)`}

\item{mode}{Character vector of modes of travel valid values TRANSIT, WALK, BICYCLE, CAR, BUS, RAIL, default CAR}

\item{date_time}{POSIXct, a date and time, defaults to current date and time}

\item{arriveBy}{Logical, Whether the trip should depart or arrive at the specified date and time, default FALSE}

\item{maxWalkDistance}{Numeric passed to OTP}

\item{walkReluctance}{Numeric passed to OTP}

\item{transferPenalty}{Numeric passed to OTP}

\item{minTransferTime}{Numeric passed to OTP}

\item{full_elevation}{Logical, should the full elevation profile be returned, defualt FALSE}
}
\description{
Get the geometry of a route from the OTP
}
