% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-batch.R
\name{otp_plan_batch}
\alias{otp_plan_batch}
\title{Get multiple routes from the OTP}
\usage{
otp_plan_batch(otpcon = NA, fromPlace = NA, toPlace = NA, ...)
}
\arguments{
\item{otpcon}{OTP connection object produced by otp_connect()}

\item{fromPlace}{Numeric matrix of two columns, Latitude/Longitude pairs}

\item{toPlace}{Numeric matrix of two columns, Latitude/Longitude pairs}

\item{...}{Other variaibles passed to otp_plan}
}
\value{
Returns a data.frame of SF POLYLINES
}
\description{
Get multiple routes from the OTP
}
