% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-setup.R
\name{otp_setup}
\alias{otp_setup}
\title{Set up an OTP instance.}
\usage{
otp_setup(otp = NULL, dir = NULL, memory = 2, router = "current",
  port = 8080, securePort = 8081, analyst = FALSE, wait = TRUE)
}
\arguments{
\item{otp}{A character string, path to the OTP .jar file}

\item{dir}{A character string, path to a directory containing the necessary files, see details}

\item{memory}{A positive integer. Amount of memory to assign to the OTP in GB, default is 2}

\item{router}{A character vector for the name of the routers, must match with contents of dir, default "current"
Only a single router is currently supported}

\item{port}{A positive integer. Optional, default is 8080.}

\item{securePort}{A positive integer. Optional, default is 8081.}

\item{analyst}{Logical. Should the analyist features be loaded? Default FALSE}

\item{wait}{Logical, Should R wait until OTP has loaded before running next line of code, default TRUE}
}
\value{
This function does not return a value to R.
If wait is TRUE R will wait until OTP is running (maximum of 5 minutes)
}
\description{
OTP is run in Java and requires Java commands to be typed into the command line.
The function allows the parameters to be defined in R and automatically passed to Java.
This function sets up a local instance of OTP, for remote versions see documentation.
}
\details{
The function assumes you have run otp_build_graph()
}
\examples{
\dontrun{
otp_setup(otp = "C:/otp/otp.jar", dir = "C:/data")
otp_setup(otp = "C:/otp/otp.jar", dir = "C:/data", memory = 5, analyst = TRUE)
}
}
