% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-setup.R
\name{otp_stop}
\alias{otp_stop}
\title{Stop and OTP Instance}
\usage{
otp_stop()
}
\description{
OTP is run in Java and requires Java commands to be typed into the command line.
The function allows the parameters to be defined in R and automatically passed to Java.
This function stops an already running OTP instance
}
\details{
The function assumes you have run otp_setup()
}
