% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/otp-surface.R
\name{otp_surface}
\alias{otp_surface}
\title{Use OTP Geo-coder to find a location}
\usage{
otp_surface(otpcon = NULL, surface = NULL, pointsset = NULL, get_data = TRUE)
}
\arguments{
\item{otpcon}{OTP connection object produced by otp_connect()}

\item{surface}{A suface list from otp_make_surface()}

\item{pointsset}{character, name of pointset}

\item{get_data}{Logical, should data be returned or just travel times.}
}
\value{
Returns a data.frame of travel times
}
\description{
Geo-coding converts a named place, such as a street name into a lng/lat pair.
}
\details{
THis function requires the analysis and pointset features to be
enabled during `otp_setup()`. Thus it will only work with OTP 1.x. For more
detail see the analyst vignette.
}
\examples{
\dontrun{
times <- otp_surface(otpcon, c(-1.17502, 50.64590), "lsoa", path_data)
}
}
