% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/json2sf.R
\name{polyline2linestring}
\alias{polyline2linestring}
\title{Convert Google Encoded Polyline into sf object}
\usage{
polyline2linestring(line)
}
\arguments{
\item{line}{character - polyline}
}
\description{
Convert Google Encoded Polyline into sf object
}
\examples{
None
}
